add "fieldtypes" package
