+++ src/main/java/com/kdgregory/geoutil/lib/kml/Feature.java
@@ -20,6 +20,8 @@ import net.sf.practicalxml.DomUtil;
 
 import com.kdgregory.geoutil.lib.internal.ObjectUtils;
 import com.kdgregory.geoutil.lib.internal.XmlUtils;
+import com.kdgregory.geoutil.lib.kml.fieldtypes.TimeSpan;
+import com.kdgregory.geoutil.lib.kml.fieldtypes.TimeStamp;
 
 
 /**
