+++ src/main/java/com/kdgregory/geoutil/lib/kml/IconStyle.java
@@ -20,6 +20,7 @@ import net.sf.practicalxml.DomUtil;
 
 import com.kdgregory.geoutil.lib.internal.ObjectUtils;
 import com.kdgregory.geoutil.lib.internal.XmlUtils;
+import com.kdgregory.geoutil.lib.kml.fieldtypes.ColorMode;
 
 
 /**
