+++ src/main/java/com/kdgregory/geoutil/lib/kml/LineString.java
@@ -27,6 +27,8 @@ import net.sf.practicalxml.DomUtil;
 
 import com.kdgregory.geoutil.lib.internal.ObjectUtils;
 import com.kdgregory.geoutil.lib.internal.XmlUtils;
+import com.kdgregory.geoutil.lib.kml.fieldtypes.AltitudeMode;
+import com.kdgregory.geoutil.lib.kml.fieldtypes.Coordinates;
 import com.kdgregory.geoutil.lib.shared.Point;
