+++ src/main/java/com/kdgregory/geoutil/lib/kml/fieldtypes/Coordinates.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.kdgregory.geoutil.lib.kml;
+package com.kdgregory.geoutil.lib.kml.fieldtypes;
 
 import java.util.ArrayList;
 import java.util.List;
