+++ src/main/java/com/kdgregory/geoutil/util/kml/GarminTrackCompare.java
@@ -22,6 +22,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.kdgregory.geoutil.lib.kml.*;
+import com.kdgregory.geoutil.lib.kml.fieldtypes.Coordinates;
 import com.kdgregory.geoutil.lib.shared.*;
