+++ src/test/java/com/kdgregory/geoutil/lib/kml/TestStyle.java
@@ -26,6 +26,7 @@ import net.sf.practicalxml.DomUtil;
 import net.sf.practicalxml.builder.XmlBuilder;
 
 import com.kdgregory.geoutil.lib.internal.XmlUtils;
+import com.kdgregory.geoutil.lib.kml.fieldtypes.ColorMode;
 
 
 public class TestStyle
