+++ src/test/java/com/kdgregory/geoutil/lib/kml/fieldtypes/TestCoordinates.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.kdgregory.geoutil.lib.kml;
+package com.kdgregory.geoutil.lib.kml.fieldtypes;
 
 import java.util.Arrays;
 import java.util.List;
@@ -20,6 +20,7 @@ import java.util.List;
 import org.junit.Test;
 import static org.junit.Assert.*;
 
+import com.kdgregory.geoutil.lib.kml.fieldtypes.Coordinates;
 import com.kdgregory.geoutil.lib.shared.Point;
